photo_vk_likes_estimation/requirements.txt
Scrapy==2.4.0
mongoengine==0.17.0
Pillow==8.0.1

flask==1.1.2
tensorflow==2.4.0
pandas==1.2.0
python-resize-image==1.1.19
image-classifiers==1.0.0b1

schedule==0.6.0
scikit-learn==0.24.0
keras_applications==1.0.8
